requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
jwt==1.3.1
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
